+++ examples/quick_start/srt_example_openvla.py
@@ -13,7 +13,9 @@ def image_qa(s, image_path, question):
 
 def single():
     state = image_qa.run(
-        image_path="images/robot.jpg", question="In: What action should the robot take to {<INSTRUCTION>}?\nOut:", max_new_tokens=7
+        image_path="images/robot.jpg",
+        question="In: What action should the robot take to {<INSTRUCTION>}?\nOut:",
+        max_new_tokens=7,
     )
     print(state["answer"], "\n")
 
@@ -48,7 +50,6 @@ if __name__ == "__main__":
         model_path="openvla/openvla-7b",
         tokenizer_path="openvla/openvla-7b",
         disable_cuda_graph=True,
-        chat_template="openvla"
     )
     sgl.set_default_backend(runtime)
     print(f"chat template: {runtime.endpoint.chat_template.name}")
